TNO-1742: Move the Generate Elasticsearch Query...
