+++ libs/npm/core/src/components/index.ts
@@ -13,6 +13,7 @@ export * from './menu';
 export * from './modal';
 export * from './navbar';
 export * from './section';
+export * from './sentiment-slider';
 export * from './show';
 export * from './spinners';
 export * from './table';
@@ -21,4 +22,3 @@ export * from './toggle-group';
 export * from './tool-bar';
 export * from './user-menu';
 export * from './view';
-export * from './sentiment-slider';
