+++ libs/npm/core/src/hooks/api-subscriber/index.ts
@@ -2,6 +2,6 @@ export * from './interfaces';
 export * from './useApiSubscriberCache';
 export * from './useApiSubscriberContents';
 export * from './useApiSubscriberMinisters';
+export * from './useApiSubscriberReports';
 export * from './useApiSubscriberSystemMessages';
 export * from './useApiSubscriberWorkOrders';
-export * from './useApiSubscriberReports';
