+++ libs/npm/core/src/utils/generateRangeForArrayField.ts
