+++ libs/npm/core/src/utils/generateSimpleQueryString.ts
