+++ libs/npm/core/src/utils/generateTerm.ts
