+++ libs/npm/core/src/utils/generateTerms.ts
