+++ libs/npm/core/src/utils/generateTermsForArrayField.ts
