+++ libs/npm/core/src/utils/index.ts
@@ -7,6 +7,16 @@ export * from './extractResponseData';
 export * from './filterEnabledOptions';
 export * from './formatIdir';
 export * from './fromQueryString';
+export * from './generateQuery';
+export * from './generateQueryForAction';
+export * from './generateQueryForActions';
+export * from './generateRangeForArrayField';
+export * from './generateRangeForDateOffset';
+export * from './generateRangeForDates';
+export * from './generateSimpleQueryString';
+export * from './generateTerm';
+export * from './generateTerms';
+export * from './generateTermsForArrayField';
 export * from './getCookie';
 export * from './getDirectoryName';
 export * from './getEnumStringOptions';
